cleanup; separates AdList into its own component
