+++ backend/server/public/src/components/frontend.jsx
@@ -1,13 +1,10 @@
 import React from "react";
-import { connect } from "react-redux";
-import { Route, Switch, withRouter } from "react-router-dom";
+import { Route, Switch } from "react-router-dom";
 import AdDetail from "components/addetail.jsx";
 import AdList from "components/adlist.jsx";
-import { newSearch } from "actions.js";
-import { debounce } from "lodash";
 import { t } from "i18n.js";
 
-export const FrontendUnconnected = () => (
+export const FrontEnd = () => (
   <div id="app">
     <div dangerouslySetInnerHTML={{ __html: t("guff") }} />
     <Switch>
@@ -19,22 +16,4 @@ export const FrontendUnconnected = () => (
   </div>
 );
 
-const throttledDispatch = debounce((dispatch, input) => {
-  dispatch(newSearch(input));
-}, 750);
-
-const Frontend = withRouter(
-  connect(
-    ({ ads, search, ad }) => ({ ads, search, ad }),
-    dispatch => ({
-      onKeyUp: e => {
-        e.preventDefault();
-        throttledDispatch(
-          dispatch,
-          e.target.value.length ? e.target.value : null
-        );
-      }
-    })
-  )(FrontendUnconnected)
-);
-export default Frontend;
+export default FrontEnd;
